add unix timeto dateTime and dateTime to unixtime operations

+++ app/src/main/java/rothkegel/com/todoapp/api/connector/utils/ToDo.kt
@@ -4,7 +4,7 @@ data class ToDo(
         var id: Int,
         var name: String,
         var description: String,
-        var expiry: Int,
+        var expiry: Long,
         var done: Boolean,
         var favourite: Boolean,
         var contacts: Array<String>?,
